adding pytest_plugin to python so other repos can access
